+++ graphql/version_resolver.go
@@ -20,7 +20,6 @@ import (
 	"github.com/evergreen-ci/evergreen/model/user"
 	"github.com/evergreen-ci/evergreen/rest/data"
 	restModel "github.com/evergreen-ci/evergreen/rest/model"
-	"github.com/evergreen-ci/gimlet"
 	"github.com/evergreen-ci/utility"
 	werrors "github.com/pkg/errors"
 	"go.mongodb.org/mongo-driver/bson"
@@ -367,28 +366,6 @@ func (r *versionResolver) TaskQuarantinedTestsSample(ctx context.Context, obj *r
 	}
 
 	return apiSamples, nil
-}
-
-func requireVersionTasksViewPermission(ctx context.Context, versionID string) error {
-	params, err := data.BuildProjectParameterMapForGraphQL(map[string]any{"versionId": versionID})
-	if err != nil {
-		return InputValidationError.Send(ctx, err.Error())
-	}
-	projectID, statusCode, err := data.GetProjectIdFromParams(ctx, params)
-	if err != nil {
-		return mapHTTPStatusToGqlError(ctx, statusCode, err)
-	}
-
-	usr := mustHaveUser(ctx)
-	if usr.HasPermission(ctx, gimlet.PermissionOpts{
-		Resource:      projectID,
-		ResourceType:  evergreen.ProjectResourceType,
-		Permission:    evergreen.PermissionTasks,
-		RequiredLevel: evergreen.TasksView.Value,
-	}) {
-		return nil
-	}
-	return Forbidden.Send(ctx, fmt.Sprintf("user '%s' does not have permission to 'view tasks' for the project '%s'", usr.Username(), projectID))
 }
 
 // Tasks is the resolver for the tasks field.
